feat(model3d): big-rig support — CPU skinning path, u16 joint indices, pose-hold, idle sway
- Model3DVertex.joints widened to uint16 (film rigs exceed 255 joints)
- CPU-skinning fallback when joint count exceeds the uniform palette:
  poses verts on CPU into a dynamic VBO, shader draws identity skin;
  scratch verts neutralized to slot 0 so the shader skin term is inert
- One-shot clips shorter than 1.5s hold their final frame (pose-library
  clips no longer flash)
- Procedural idle sway (bob + rock) while idling — gives life to packs
  whose clips are static poses
- Ellie pack (Blender Studio, Sprite Fright): 329-joint rig, 12 clips,
  bust framing, flat-color palette export (procedural Blender materials
  don't survive glTF export); conversion script for reproducibility

+++ src/model3d/Model3DTypes.h
@@ -18,7 +18,7 @@ struct Model3DVertex {
     float px, py, pz;
     float nx, ny, nz;
     float u, v;
-    uint8_t joints[4];
+    uint16_t joints[4];         // u16: rigs can exceed 255 joints (film rigs)
     float weights[4];
 };
